Add CIFAR100 dataset

+++ prepare_dataset.py
@@ -2,6 +2,10 @@ import requests
 from tqdm import tqdm
 import os
 import zipfile
+import numpy as np
+import torchvision
+import torchvision.transforms as transforms
+import h5py
 
 def download_dataset(dset_name, link, save_dir):
     os.makedirs(save_dir, exist_ok=True)
@@ -20,27 +24,71 @@ def download_dataset(dset_name, link, save_dir):
     return os.path.join(save_dir, f'{dset_name}.zip')
 
 
+def cifar_to_h5py(dset, filename, num_classes = 100, num_tasks = 20, val_split = 0.2, isTrain = True):
+    classes_shuff = np.random.permutation(np.arange(num_classes))
+    label_to_task = {ii:i%num_tasks for i,ii in enumerate(classes_shuff)}
+
+    train_dset = [[] for _ in range(num_tasks)]
+    for item in dset:
+        lab = item[1]
+        task_num = label_to_task[lab]
+        train_dset[task_num].append(item)
+    if isTrain:
+        # Need to create the val file
+        f_val = h5py.File(filename.replace('_train', '_val'),'w')
+    f  = h5py.File(filename,'w')
+    for task in range(20):
+        X = np.stack([x[0] for x in train_dset[task]])
+        Y = np.stack([x[1] for x in train_dset[task]])
+        if isTrain:
+            indices = np.arange(X.shape[0])
+            np.random.shuffle(indices)
+            size = int(val_split*X.shape[0])
+            X_val = X[indices[:size]]
+            Y_val = Y[indices[:size]]
+            X = X[indices[size:]]
+            Y = Y[indices[size:]]
+            gr_val = f_val.create_group(str(task))
+            gr_val.create_dataset('X',data=X_val)
+            gr_val.create_dataset('Y',data=Y_val)
+
+        gr = f.create_group(str(task))
+        gr.create_dataset('X',data=X)
+        gr.create_dataset('Y',data=Y)
+    f.close()
+    if isTrain:
+        f_val.close()
+
 def prepare(dset, save_dir = 'datasets'):
-    if not dset['link']: dset['link'] = 'https://github.com/gmshroff/metaLearning2022/raw/main/data/{}.zip'.format(dset['filename'])
-    zip_file = download_dataset(dset['name'], dset['link'], save_dir)
-    with zipfile.ZipFile(zip_file) as zf:
-        zf.extractall(save_dir)
-    os.remove(os.path.join(save_dir, f'{dset["name"]}.zip'))
-    return os.path.join(save_dir, dset['filename'])
+    if dset['name'] == 'cifar100':
+        tr_set = torchvision.datasets.CIFAR100(root='./cifar100', train=True, download=True, transform=transforms.ToTensor())
+        cifar_to_h5py(tr_set, os.path.join(save_dir, 'cifar100_train.h5'))
+        ts_set = torchvision.datasets.CIFAR100(root='./cifar100', train=False, download=True, transform=transforms.ToTensor())
+        cifar_to_h5py(ts_set, os.path.join(save_dir, 'cifar100_test.h5'), isTrain = False)
+    else:
+        if not dset['link']: dset['link'] = 'https://github.com/gmshroff/metaLearning2022/raw/main/data/{}.zip'.format(dset['filename'])
+        zip_file = download_dataset(dset['name'], dset['link'], save_dir)
+        with zipfile.ZipFile(zip_file) as zf:
+            zf.extractall(save_dir)
+        os.remove(os.path.join(save_dir, f'{dset["name"]}.zip'))
+        return os.path.join(save_dir, dset['filename'])
 
 if __name__ == '__main__':
 
     datasets = {
-        'image_data': {
-            'name': 'image_data',
-            'filename': 'image_data.pickle',
-            'link' : '',
-        },
-        'image_grid_data': {
-            'name': 'image_grid_data',
-            'filename': 'image_grid_data.pickle',
-            'link' : '',
-        },
+        # 'image_data': {
+        #     'name': 'image_data',
+        #     'filename': 'image_data.pickle',
+        #     'link' : '',
+        # },
+        # 'image_grid_data': {
+        #     'name': 'image_grid_data',
+        #     'filename': 'image_grid_data.pickle',
+        #     'link' : '',
+        # },
+        'cifar100' : {
+            'name': 'cifar100',
+        }
     }
 
     for dset in datasets:
